Added routes to routes.js
